Added a -n option to bootstrap command to specify a deployment name and clarified HostsPool CLI
the -n option allows user to invoke [yorc bootstrap -n somename] in order to save the config done using CLI invite after into a special deployment name.
Case where a duplicate name already exists is not yet handled.

When running hostspool configuration, the [yorc bootstrap] command would ask user to define labels for the host in the form of a key/value couple.
A public_address key has to be defined and associated with a value. The message were not clearly explicit. Now the CLI asks for the public_address
value first and AFTER loops on the other possibilities for the host.