Change dep status to UNDEPLOYMENT_IN_PROGRESS asap
That means that instead of changing status when the corresponding task
is processed by  a worker, do it as soon as the task is registered.
This way the user sees its requested as acknowledged.

Also corrected some issues slurm allocation (potential goroutines leaks)

And a AnotherLivingTaskAlreadyExistsError not properly detected
and retruned as 400 in the REST server.