Fix vulnerability GHSA-8vhw-qv5r-38h5
Refactored the way we get and validate private keys in order to
check if the key is a valid file path containing a valid key or if
it is a valid key content.

Now all provisioning packages use this method that is now in the sshutil
package.

As some kind of side effect we implemented the ability to support multiple
private keys when the ssh-agent is enabled.